Update README.md
This project studies the economic results of 173 undergraduate majors using the FiveThirtyEight “College Majors” dataset. It looks at the relationship between median income and unemployment rates. The aim is to show the difference between technical fields and other areas of study. The analysis is done in R, and ggplot2 is used to create the graphs.